+++ src/Service/InvoiceGenerator/PdfGeneratorException.php
@@ -1,0 +1,7 @@
+<?php
+namespace Psys\OrderInvoiceBundle\Service\InvoiceGenerator;
+
+/**
+ * Generic exception thrown when PDF generation fails.
+ */
+class PdfGeneratorException extends \RuntimeException {}